+++ pyuvsim/tests/test_mpi.py
@@ -23,7 +23,7 @@ import pyradiosky
 
 @pytest.fixture(scope='module', autouse=True)
 def start_mpi():
-    mpi.start_mpi()
+    mpi.start_mpi(False)
 
 
 @pytest.fixture(scope='module')
@@ -50,6 +50,7 @@ def test_mpi_version():
     assert MPI.VERSION == 3
 
 
+@pytest.mark.parallel(2)
 def test_mpi_funcs():
     assert mpi.get_rank() == MPI.COMM_WORLD.rank
     assert mpi.get_Npus() == MPI.COMM_WORLD.size
@@ -96,11 +97,15 @@ def test_mem_usage():
     assert np.isclose(change, incsize / 2**30, atol=5e-2)
 
 
+@pytest.mark.parallel(4)
 def test_mpi_counter():
+    # Warning -- This test has been flaky in the past.
+    mpi.start_mpi()
     count = mpi.Counter()
     N = 20
     for i in range(N):
         count.next()
+    mpi.world_comm.Barrier()
     assert count.current_value() == N * mpi.world_comm.size
     count.free()
 
@@ -129,6 +134,7 @@ def test_big_gather(MAX_BYTES, fake_tasks):
             assert len(split_info['ranges']) > 1
 
 
+@pytest.mark.parallel(3)
 @pytest.mark.parametrize('MAX_BYTES', [mpi.INT_MAX, 100])
 def test_big_bcast(MAX_BYTES, fake_tasks):
 
@@ -154,6 +160,7 @@ def test_big_bcast(MAX_BYTES, fake_tasks):
             assert len(split_info['ranges']) > 1
 
 
+@pytest.mark.parallel(3)
 def test_big_bcast_gather_loop(fake_tasks):
 
     objs = fake_tasks
@@ -165,6 +172,7 @@ def test_big_bcast_gather_loop(fake_tasks):
         assert broadcast == gathered[0]
 
 
+@pytest.mark.parallel(3)
 def test_sharedmem_bcast_with_quantities():
     # Use mpi.quantity_shared_bcast and check returned objects.
 
@@ -179,6 +187,7 @@ def test_sharedmem_bcast_with_quantities():
         assert np.all(freq_return.to("MHz") == freqs.to("MHz"))
 
 
+@pytest.mark.parallel(3)
 def test_skymodeldata_share():
     # Test the SkyModelData share method.
     sky = pyradiosky.SkyModel(